Add middleware for recurring URL computations
When we match a url, we almost always have a matching `label_id`, `version`,
etc. which we then pass around again and again. This patch removes some of
that repetition by setting these arguments (and a computed `cfr_part`
argument, when applicable) on an attribute of the request object. We don't
currently use this anywhere, but it should allow us to clean up some
repetition.